Clean up and organize imports
