+++ stograde/common/__init__.py
@@ -4,7 +4,6 @@ from .check_submit_date import check_dates
 from .dirsize import dirsize
 from .find_unmerged_branches_in_cwd import find_unmerged_branches_in_cwd
 from .flatten import flatten
-from .group_by import group_by
 from .parse_commit_msg_for_assignments import parse_commit_msg_for_assignments
 from .pipe import pipe
 from .run import run
