+++ stograde/common/group_by.py
@@ -1,12 +1,10 @@
 from itertools import groupby
 from types import LambdaType
-from typing import Iterable, List, Any, Tuple
-
-__all__ = ['group_by']
-
-from typing import Generator
+from typing import Any, Generator, Iterable, List, Tuple
 
 from ..process_assignment.record_result import RecordResult
+
+__all__ = ['group_by']
 
 
 def group_by(iterable: Iterable[RecordResult],
