+++ stograde/formatters/base.py
@@ -25,9 +25,9 @@ def format_collected_data(student_results: List[StudentResult],
         results.extend(student.results)
 
     if group_by == 'assignment':
-        grouped_records = group(results, lambda rec: rec.spec_id)
+        grouped_records = group.group_by(results, lambda rec: rec.spec_id)
     elif group_by == 'student':
-        grouped_records = group(results, lambda rec: rec.student)
+        grouped_records = group.group_by(results, lambda rec: rec.student)
     else:
         # not entirely sure what this'll do
         grouped_records = results
