+++ stograde/process_file/__init__.py
@@ -1,2 +1,1 @@
-from .file_result import FileResult
 from .process_file import process_file
