+++ stograde/specs/__init__.py
@@ -1,7 +1,3 @@
 from .cache import delete_cache
-from .file_options import FileOptions
 from .load import load_all_specs, load_some_specs
-from .spec import Spec
-from .spec_file import SpecFile
-from .supporting_file import SupportingFile
 from .util import check_architecture, check_dependencies, get_filenames
