+++ stograde/specs/cache.py
@@ -1,11 +1,10 @@
-import shutil
-import sys
-from logging import debug
-import json
 import copy
+import json
+from logging import debug
 import os
 from pathlib import Path
-
+import shutil
+import sys
 import yaml
