+++ stograde/student/stash.py
@@ -1,7 +1,6 @@
 import logging
 
-from stograde.common import chdir
-from stograde.common import run
+from ..common import chdir, run
 
 
 def stash(student: str, no_repo_update: bool):
