+++ stograde/toolkit/process_student.py
@@ -1,7 +1,6 @@
 from typing import List
 
-from ..process_assignment import find_unmerged_branches
-from ..specs import Spec
+from ..specs.spec import Spec
 from ..student import analyze, checkout_date, clone_student, pull, record, remove, reset, stash
 from ..student.student_result import StudentResult
